wip: PR revision
